Button state styling, new model for text and button fields
