+++ src/fields/ButtonElement.jsx
@@ -1,0 +1,162 @@
+import React, { useState } from 'react';
+import Button from 'react-bootstrap/Button';
+import Spinner from 'react-bootstrap/Spinner';
+import Delta from 'quill-delta';
+
+import { adjustColor, alignmentMap } from '../utils/formHelperFunctions';
+import generateNodes from './textNodes';
+
+function ButtonElement({
+    field,
+    fieldValues,
+    displaySteps,
+    addRepeatedRow,
+    removeRepeatedRow,
+    submit
+}) {
+    const [showSpinner, setShowSpinner] = useState(false);
+
+    const elementKey = field.text;
+    const repeat = field.repeat || 0;
+
+    const delta = new Delta(field.text_formatted);
+
+    const hoverStyles = {};
+    const selectedStyles = {};
+    if (field.hover_border_color)
+        hoverStyles.borderColor = `#${field.hover_border_color} !important`;
+    if (field.hover_background_color)
+        hoverStyles.backgroundColor = `#${field.hover_background_color} !important`;
+    if (field.hover_font_color)
+        hoverStyles.color = `#${field.hover_font_color} !important`;
+    if (field.selected_border_color)
+        selectedStyles.borderColor = `#${field.selected_border_color} !important`;
+    if (field.selected_background_color)
+        selectedStyles.backgroundColor = `#${field.selected_background_color} !important`;
+    if (field.selected_font_color)
+        selectedStyles.color = `#${field.selected_font_color} !important`;
+
+    const nodes = generateNodes({
+        delta,
+        fieldValues,
+        field,
+        submit,
+        repeat,
+        elementKey
+    });
+
+    async function buttonOnClick() {
+        if (displaySteps || field.link === 'none') {
+            return;
+        }
+
+        if (field.link === 'add_repeated_row') {
+            addRepeatedRow();
+            return;
+        }
+
+        if (field.link === 'remove_repeated_row') {
+            removeRepeatedRow(field.repeat);
+            return;
+        }
+
+        if (field.link === 'submit') {
+            setShowSpinner(true);
+        }
+
+        await submit(
+            field.link === 'submit',
+            {
+                elementType: 'button',
+                elementKeys: [elementKey],
+                trigger: 'click'
+            },
+            repeat
+        );
+        setShowSpinner(false);
+    }
+
+    return (
+        <div
+            css={{
+                gridColumnStart: field.column_index + 1,
+                gridRowStart: field.row_index + 1,
+                gridColumnEnd: field.column_index_end + 2,
+                gridRowEnd: field.row_index_end + 2,
+                paddingBottom: `${field.padding_bottom}px`,
+                paddingTop: `${field.padding_top}px`,
+                paddingLeft: `${field.padding_left}px`,
+                paddingRight: `${field.padding_right}px`,
+                display: 'flex',
+                flexDirection: 'column',
+                alignItems: alignmentMap[field.layout],
+                textAlign: field.layout,
+                justifyContent: field.vertical_layout
+            }}
+        >
+            <Button
+                key={elementKey}
+                style={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center',
+                    cursor: field.link === 'none' ? 'default' : 'pointer',
+                    borderRadius: `${field.border_radius}px`,
+                    borderColor: `#${field.border_color}`,
+                    backgroundColor: `#${field.button_color}`,
+                    boxShadow: 'none',
+                    height: `${field.button_height}${field.button_height_unit}`,
+                    width: `${field.button_width}${field.button_width_unit}`,
+                    maxWidth: '100%'
+                }}
+                css={{
+                    '&:active': selectedStyles,
+                    '&:hover': hoverStyles,
+                    '&:disabled': {
+                        cursor: 'default !important'
+                    },
+                    '&:hover:enabled':
+                        field.link !== 'none'
+                            ? {
+                                  backgroundColor: `${adjustColor(
+                                      field.button_color,
+                                      -30
+                                  )} !important`,
+                                  borderColor: `${adjustColor(
+                                      field.button_color,
+                                      -30
+                                  )} !important`,
+                                  transition: 'background 0.3s !important'
+                              }
+                            : {}
+                }}
+                disabled={field.link === 'none'}
+                onClick={buttonOnClick}
+            >
+                <div style={{ display: 'flex', position: 'relative' }}>
+                    {nodes}
+                    {showSpinner && (
+                        <Spinner
+                            animation='border'
+                            style={{
+                                color: 'white',
+                                position: 'absolute',
+                                right: `-${field.button_height}${field.button_height_unit}`,
+                                width: `${field.button_height / 2}${
+                                    field.button_height_unit
+                                }`,
+                                height: `${field.button_height / 2}${
+                                    field.button_height_unit
+                                }`,
+                                border: '0.2em solid currentColor',
+                                borderRightColor: 'transparent'
+                            }}
+                        />
+                    )}
+                </div>
+            </Button>
+        </div>
+    );
+}
+
+export default ButtonElement;
